create example service api

+++ services/src/main/scala/com/gemfire/models/AggregatedPosition.scala
@@ -1,5 +1,5 @@
 package com.gemfire.models
 
 case class AggregatedPosition(positions: Seq[DerivedPosition]) {
-  def balance: BigInt = positions.map(p => BigInt(p.value)).sum
+  def balance: BigDecimal = positions.map(p => BigDecimal(p.value)).sum
 }
